Changed the sorting method from bubblesort to mergesort

+++ include/ft_nm.h
@@ -6,7 +6,7 @@
 /*   By: franmart <franmart@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/28 16:36:10 by franmart          #+#    #+#             */
-/*   Updated: 2024/10/22 21:54:22 by franmart         ###   ########.fr       */
+/*   Updated: 2024/10/24 21:45:54 by franmart         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,7 +51,7 @@ void	read_symbols_elf32(t_ELF32_symbol *symtab, uint32_t n_sym, uint16_t n_sh,
 void	print_elf32_symbol(t_ELF32_symbol *sym, uint16_t n_sh, t_ELF32_section_header *sh,
 							char *strtab);
 char    get_symbol_class_elf32(t_ELF32_symbol *sym, t_ELF32_section_header *sh);
-t_ELF32_symbol	**sort_elf32_list(t_ELF32_symbol **list, uint32_t len, char *str_table);
+t_ELF32_symbol	**sort_elf32_list(t_ELF32_symbol **list, uint64_t len, char *str_table);
 
 
 /* file_utils.c */
